historical-quotes/pyproject.toml
[project]
name = "historical-quotes"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.12.0",
    "fastavro>=1.12.1",
    "python-dotenv>=1.1.1",
]
